rename reorganize some classes add log messages
