--- src/main/java/tic_tac_toe/service/package-info.java
@@ -1,1 +1,0 @@
-package tic_tac_toe.service;
