+++ src/main/java/tic_tac_toe/service/validate/ValidationService.java
@@ -1,4 +1,4 @@
-package tic_tac_toe.validate;
+package tic_tac_toe.service.validate;
 
 import tic_tac_toe.pojo.GamePlayRequest;
 import tic_tac_toe.pojo.NewGameRequest;
